+++ vcfpy/header.py
@@ -260,7 +260,7 @@ class Header:
     """Represent header of VCF file
 
     While this class allows mutating records, it should not be changed once it
-    has been assigned to
+    has been assigned to a writer
 
     This class provides function for adding lines to a header and updating the
     supporting index data structures.  There is no explicit API for removing