+++ HISTORY.rst
@@ -5,6 +5,8 @@ History
 HEAD
 ----
 
+* Using ``samples`` attribute from ``Header`` in ``Reader`` and ``Writer`` instead of passing explicitely
+
 0.9.0 (2017-02-26)
 ------------------